commands: move rebase code to rebase.rs
